+++ packages/rnv/src/platformTools/webos/index.js
@@ -9,7 +9,6 @@ import {
     getAppFolder,
     getAppVersion,
     getAppTitle,
-    writeCleanFile,
     getAppId,
     getAppTemplateFolder,
     getConfigProp,
@@ -37,7 +36,7 @@ import {
     CLI_WEBOS_ARES_NOVACOM,
     CLI_WEBOS_ARES_SETUP_DEVICE
 } from '../../constants';
-import { getRealPath } from '../../systemTools/fileutils';
+import { getRealPath, writeCleanFile } from '../../systemTools/fileutils';
 import { buildWeb, configureCoreWebProject } from '../web';
 import { rnvStart } from '../runner';
 import Config from '../../config';
@@ -291,22 +290,21 @@ const runWebOS = async (c, platform, target) => {
     }
 };
 
-const buildWebOSProject = (c, platform) => new Promise((resolve, reject) => {
+const buildWebOSProject = async (c, platform) => {
     logTask(`buildWebOSProject:${platform}`);
 
-    const tDir = path.join(getAppFolder(c, platform), 'public');
-    const tOut = path.join(getAppFolder(c, platform), 'output');
+    await buildWeb(c, platform);
 
-    buildWeb(c, platform)
-        .then(() => execCLI(c, CLI_WEBOS_ARES_PACKAGE, `-o ${tOut} ${tDir} -n`))
-        .then(() => {
-            logSuccess(
-                `Your IPK package is located in ${chalk.white(tOut)} .`
-            );
-            return resolve();
-        })
-        .catch(reject);
-});
+    if (!c.program.hosted) {
+        const tDir = path.join(getAppFolder(c, platform), 'public');
+        const tOut = path.join(getAppFolder(c, platform), 'output');
+        await execCLI(c, CLI_WEBOS_ARES_PACKAGE, `-o ${tOut} ${tDir} -n`);
+
+        logSuccess(
+            `Your IPK package is located in ${chalk.white(tOut)} .`
+        );
+    }
+};
 
 const configureWebOSProject = async (c, platform) => {
     logTask('configureWebOSProject');
@@ -341,7 +339,7 @@ const configureProject = (c, platform) => new Promise((resolve, reject) => {
                 pattern: '{{APP_VERSION}}',
                 override: semver.coerce(getAppVersion(c, platform))
             }
-        ]
+        ], null, c
     );
 
     resolve();