+++ packages/rnv/src/projectTools/projectParser.js
@@ -7,7 +7,8 @@ import {
     getAppSubFolder,
     getBuildsFolder,
     areNodeModulesInstalled,
-    getConfigProp
+    getConfigProp,
+    getTimestampPathsConfig
 } from '../common';
 import { doResolve } from '../resolve';
 import {
@@ -15,7 +16,8 @@ import {
     copyFolderContentsRecursiveSync,
     copyFileSync,
     mkdirSync,
-    writeFileSync
+    writeFileSync,
+    fsWriteFileSync
 } from '../systemTools/fileutils';
 import { isPlatformActive } from '../platformTools';
 import { npmInstall } from '../systemTools/exec';
@@ -51,7 +53,7 @@ export const checkAndCreateProjectPackage = c => new Promise((resolve) => {
         };
         pkgJson.devDependencies[templateName] = rnvVersion;
         const pkgJsonStringClean = JSON.stringify(pkgJson, null, 2);
-        fs.writeFileSync(c.paths.project.package, pkgJsonStringClean);
+        fsWriteFileSync(c.paths.project.package, pkgJsonStringClean);
     }
 
     loadFile(c.files.project, c.paths.project, 'package');
@@ -161,11 +163,11 @@ export const copyRuntimeAssets = c => new Promise((resolve, reject) => {
 
         if (existingFileContents !== fontsObj) {
             logDebug('newFontsJsFile');
-            fs.writeFileSync(fontJsPath, fontsObj);
+            fsWriteFileSync(fontJsPath, fontsObj);
         }
     } else {
         logDebug('newFontsJsFile');
-        fs.writeFileSync(fontJsPath, fontsObj);
+        fsWriteFileSync(fontJsPath, fontsObj);
     }
 
     const supportFiles = path.resolve(c.paths.rnv.dir, 'supportFiles');
@@ -274,6 +276,8 @@ export const copyAssetsFolder = async (c, platform, customFn) => {
         ASSET_PATH_ALIASES[platform]
     );
 
+    const tsPathsConfig = getTimestampPathsConfig(c, platform);
+
     // FOLDER MERGERS FROM APP CONFIG + EXTEND
     if (c.paths.appConfig.dirs) {
         const hasAssetFolder = c.paths.appConfig.dirs.filter(v => fs.existsSync(path.join(v, `assets/${platform}`))).length;
@@ -286,7 +290,7 @@ export const copyAssetsFolder = async (c, platform, customFn) => {
         }
         c.paths.appConfig.dirs.forEach((v) => {
             const sourcePath = path.join(v, `assets/${platform}`);
-            copyFolderContentsRecursiveSync(sourcePath, destPath);
+            copyFolderContentsRecursiveSync(sourcePath, destPath, true, false, false, null, tsPathsConfig);
         });
     } else {
         const sourcePath = path.join(
@@ -296,7 +300,7 @@ export const copyAssetsFolder = async (c, platform, customFn) => {
         if (!fs.existsSync(sourcePath)) {
             await generateDefaultAssets(c, platform, sourcePath);
         }
-        copyFolderContentsRecursiveSync(sourcePath, destPath);
+        copyFolderContentsRecursiveSync(sourcePath, destPath, true, false, false, null, tsPathsConfig);
     }
 };
 
@@ -326,6 +330,7 @@ export const copyBuildsFolder = (c, platform) => new Promise((resolve, reject) =
     if (!isPlatformActive(c, platform, resolve)) return;
 
     const destPath = path.join(getAppFolder(c, platform));
+    const tsPathsConfig = getTimestampPathsConfig(c, platform);
 
     const configPropsInject = [];
     INJECTABLE_CONFIG_PROPS.forEach((v) => {
@@ -342,7 +347,7 @@ export const copyBuildsFolder = (c, platform) => new Promise((resolve, reject) =
         platform,
         c.paths.project.projectConfig.dir
     );
-    copyFolderContentsRecursiveSync(sourcePath1, destPath, true, false, false, configPropsInject);
+    copyFolderContentsRecursiveSync(sourcePath1, destPath, true, false, false, configPropsInject, tsPathsConfig);
 
     // FOLDER MERGERS PROJECT CONFIG (PRIVATE)
     const sourcePath1sec = getBuildsFolder(
@@ -350,7 +355,7 @@ export const copyBuildsFolder = (c, platform) => new Promise((resolve, reject) =
         platform,
         c.paths.workspace.project.projectConfig.dir
     );
-    copyFolderContentsRecursiveSync(sourcePath1sec, destPath, true, false, false, configPropsInject);
+    copyFolderContentsRecursiveSync(sourcePath1sec, destPath, true, false, false, configPropsInject, tsPathsConfig);
 
     if (WEB_HOSTED_PLATFORMS.includes(platform)) {
         // FOLDER MERGERS _SHARED
@@ -372,14 +377,14 @@ export const copyBuildsFolder = (c, platform) => new Promise((resolve, reject) =
             copyFolderContentsRecursiveSync(
                 sourceV,
                 destPath,
-                true, false, false, configPropsInject
+                true, false, false, configPropsInject, tsPathsConfig
             );
         });
     } else {
         copyFolderContentsRecursiveSync(
             getBuildsFolder(c, platform, c.paths.appConfig.dir),
             destPath,
-            true, false, false, configPropsInject
+            true, false, false, configPropsInject, tsPathsConfig
         );
     }
 
@@ -389,7 +394,7 @@ export const copyBuildsFolder = (c, platform) => new Promise((resolve, reject) =
         platform,
         c.paths.workspace.appConfig.dir
     );
-    copyFolderContentsRecursiveSync(sourcePath0sec, destPath, true, false, false, configPropsInject);
+    copyFolderContentsRecursiveSync(sourcePath0sec, destPath, true, false, false, configPropsInject, tsPathsConfig);
 
     copyTemplatePluginsSync(c, platform);
 
@@ -426,12 +431,9 @@ export const upgradeProjectDependencies = (c, version) => {
 };
 
 export const configureNodeModules = c => new Promise((resolve, reject) => {
-    logTask('configureNodeModules');
+    logTask(`configureNodeModules:${c._requiresNpmInstall}:${c.runtime.skipPackageUpdate}`);
     // Check node_modules
-    if (
-        !areNodeModulesInstalled()
-            || (c._requiresNpmInstall && !c.runtime.skipPackageUpdate)
-    ) {
+    if (!areNodeModulesInstalled() || (c._requiresNpmInstall && !c.runtime.skipPackageUpdate)) {
         if (!areNodeModulesInstalled()) {
             logWarning(
                 `Looks like your node_modules folder is missing! Let's run ${chalk.white(